+++ modules/@ergonode/activity-logs/src/pages/activity-logs/index.vue
@@ -18,7 +18,7 @@ import Page from '@Core/components/Layout/Page';
 import HorizontalRoutingTabBar from '@Core/components/TabBar/Routing/HorizontalRoutingTabBar';
 import TitleBar from '@Core/components/TitleBar/TitleBar';
 import beforeLeavePageMixin from '@Core/mixins/page/beforeLeavePageMixin';
-import pageMixin from '@Core/mixins/page/pageMixin';
+import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
 
 export default {
     name: 'UsersTabs',
@@ -29,7 +29,7 @@ export default {
     },
     mixins: [
         beforeLeavePageMixin,
-        pageMixin,
+        asyncTabsMixin,
     ],
 };
 </script>
