+++ modules/@ergonode/attributes/src/pages/attribute-groups/index.vue
@@ -48,7 +48,7 @@ import {
 } from '@Core/defaults/theme';
 import gridModalMixin from '@Core/mixins/modals/gridModalMixin';
 import beforeLeavePageMixin from '@Core/mixins/page/beforeLeavePageMixin';
-import pageMixin from '@Core/mixins/page/pageMixin';
+import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
 
 export default {
     name: 'AttributeTabs',
@@ -63,7 +63,7 @@ export default {
     mixins: [
         gridModalMixin,
         beforeLeavePageMixin,
-        pageMixin,
+        asyncTabsMixin,
     ],
     computed: {
         isAllowedToCreate() {
