+++ modules/@ergonode/categories/src/components/Pages/CategoryPage.vue
@@ -35,11 +35,13 @@
 <script>
 import PRIVILEGES from '@Categories/config/privileges';
 import editPageMixin from '@Core/mixins/page/editPageMixin';
+import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
 
 export default {
     name: 'CategoryPage',
     mixins: [
         editPageMixin,
+        asyncTabsMixin,
     ],
     computed: {
         isReadOnly() {
