+++ modules/@ergonode/category-trees/src/components/Pages/CategoryTreePage.vue
@@ -34,12 +34,14 @@
 
 <script>
 import editPageMixin from '@Core/mixins/page/editPageMixin';
+import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
 import PRIVILEGES from '@Trees/config/privileges';
 
 export default {
     name: 'CategoryTreePage',
     mixins: [
         editPageMixin,
+        asyncTabsMixin,
     ],
     computed: {
         isReadOnly() {
