+++ modules/@ergonode/collections/src/components/Pages/CollectionPage.vue
@@ -35,11 +35,13 @@
 <script>
 import PRIVILEGES from '@Collections/config/privileges';
 import editPageMixin from '@Core/mixins/page/editPageMixin';
+import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
 
 export default {
     name: 'CollectionPage',
     mixins: [
         editPageMixin,
+        asyncTabsMixin,
     ],
     computed: {
         isUserAllowedToDelete() {
