+++ modules/@ergonode/core/src/mixins/tab/asyncTabsMixin.js
@@ -1,0 +1,33 @@
+/*
+ * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
+ * See LICENSE for license details.
+ */
+import {
+    getNestedTabRoutes,
+} from '@Core/models/navigation/tabs';
+
+export default {
+    data() {
+        return {
+            asyncTabs: null,
+        };
+    },
+    watch: {
+        $route: {
+            immediate: true,
+            async handler() {
+                const tmpTabs = getNestedTabRoutes({
+                    hasAccess: this.$hasAccess,
+                    routes: this.$router.options.routes,
+                    route: this.$route,
+                });
+                const tabs = await this.$extendMethods('@Core/pages/tabs', {
+                    $this: this,
+                    tabs: tmpTabs,
+                });
+
+                this.asyncTabs = tabs.length ? Array.from(new Set([].concat(...tabs))) : tmpTabs;
+            },
+        },
+    },
+};
