+++ modules/@ergonode/product-statuses/src/components/Pages/ProductStatusPage.vue
@@ -35,12 +35,14 @@
 
 <script>
 import editPageMixin from '@Core/mixins/page/editPageMixin';
+import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
 import PRIVILEGES from '@Transitions/config/privileges';
 
 export default {
     name: 'ProductStatusPage',
     mixins: [
         editPageMixin,
+        asyncTabsMixin,
     ],
     computed: {
         isReadOnly() {
