+++ modules/@ergonode/product-templates/src/components/Pages/TemplatePage.vue
@@ -35,12 +35,14 @@
 
 <script>
 import editPageMixin from '@Core/mixins/page/editPageMixin';
+import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
 import PRIVILEGES from '@Templates/config/privileges';
 
 export default {
     name: 'TemplatePage',
     mixins: [
         editPageMixin,
+        asyncTabsMixin,
     ],
     computed: {
         isReadOnly() {
