+++ modules/@ergonode/products/src/components/Pages/ProductPage.vue
@@ -51,6 +51,11 @@ export default {
     mixins: [
         editPageMixin,
     ],
+    data() {
+        return {
+            asyncTabs: null,
+        };
+    },
     computed: {
         ...mapState('product', [
             'type',
@@ -70,23 +75,24 @@ export default {
             return this.$isReadOnly(PRIVILEGES.PRODUCT.namespace);
         },
     },
-    asyncComputed: {
-        async asyncTabs() {
-            const tmpTabs = getNestedTabRoutes({
-                hasAccess: this.$hasAccess,
-                routes: this.$router.options.routes,
-                route: this.$route,
-            });
-            const type = getKeyByValue(this.productTypes, this.type);
-            const tabs = await this.$extendMethods('@Core/pages/tabs', {
-                $this: this,
-                type,
-                tabs: tmpTabs,
-            });
+    watch: {
+        $route: {
+            immediate: true,
+            async handler() {
+                const tmpTabs = getNestedTabRoutes({
+                    hasAccess: this.$hasAccess,
+                    routes: this.$router.options.routes,
+                    route: this.$route,
+                });
+                const type = getKeyByValue(this.productTypes, this.type);
+                const tabs = await this.$extendMethods('@Core/pages/tabs', {
+                    $this: this,
+                    type,
+                    tabs: tmpTabs,
+                });
 
-            return tabs.length ? [
-                ...new Set([].concat(...tabs)),
-            ] : tmpTabs;
+                this.asyncTabs = tabs.length ? Array.from(new Set([].concat(...tabs))) : tmpTabs;
+            },
         },
     },
 };
