+++ modules/@ergonode/products/src/extends/components/Tabs/ProductVariantsTab.vue
@@ -101,11 +101,9 @@ export default {
                 key,
             }) => key);
         const filteredProductTypes = await app.$extendMethods('@Products/components/Tabs/ProductVariantsTab/filteredProductTypes', {
-            $this: this,
+            $this: app,
         });
-        const productTypes = [
-            ...new Set([].concat(...filteredProductTypes)),
-        ].join(',');
+        const productTypes = Array.from(new Set([].concat(...filteredProductTypes))).join(',');
         const defaultColumns = [
             'esa_default_image',
             'esa_default_label',
@@ -316,9 +314,7 @@ export default {
             const filteredProductTypes = await this.$extendMethods('@Products/components/Tabs/ProductVariantsTab/filteredProductTypes', {
                 $this: this,
             });
-            const productTypes = [
-                ...new Set([].concat(...filteredProductTypes)),
-            ].join(',');
+            const productTypes = Array.from(new Set([].concat(...filteredProductTypes))).join(',');
             const defaultColumns = [
                 'esa_default_image',
                 'esa_default_label',
