+++ modules/@ergonode/segments/src/components/Pages/SegmentPage.vue
@@ -35,12 +35,14 @@
 
 <script>
 import editPageMixin from '@Core/mixins/page/editPageMixin';
+import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
 import PRIVILEGES from '@Segments/config/privileges';
 
 export default {
     name: 'SegmentPage',
     mixins: [
         editPageMixin,
+        asyncTabsMixin,
     ],
     computed: {
         isReadOnly() {
