+++ modules/@ergonode/users/src/components/Pages/UserRolesPage.vue
@@ -34,12 +34,14 @@
 
 <script>
 import editPageMixin from '@Core/mixins/page/editPageMixin';
+import asyncTabsMixin from '@Core/mixins/tab/asyncTabsMixin';
 import PRIVILEGES from '@Users/config/privileges';
 
 export default {
     name: 'UserRolesPage',
     mixins: [
         editPageMixin,
+        asyncTabsMixin,
     ],
     computed: {
         isReadOnly() {
